Rename constants.js to exceptions.js
