+++ lib/exceptions.js
