+++ lib/opFns.js
@@ -3,10 +3,10 @@ const async = require('async')
 const fees = require('ethereum-common')
 const utils = require('ethereumjs-util')
 const BN = utils.BN
-const constants = require('./constants.js')
+const exceptions = require('./exceptions.js')
 const logTable = require('./logTable.js')
-const ERROR = constants.ERROR
-const VmError = constants.VmError
+const ERROR = exceptions.ERROR
+const VmError = exceptions.VmError
 const MAX_INT = 9007199254740991
 
 // the opcode functions
