+++ lib/runCall.js
@@ -3,9 +3,9 @@ const async = require('async')
 const ethUtil = require('ethereumjs-util')
 const BN = ethUtil.BN
 const fees = require('ethereum-common')
-const constants = require('./constants.js')
+const exceptions = require('./exceptions.js')
 
-const ERROR = constants.ERROR
+const ERROR = exceptions.ERROR
 
 /**
  * runs a CALL operation
