+++ lib/runCode.js
@@ -17,12 +17,12 @@ const utils = require('ethereumjs-util')
 const Block = require('ethereumjs-block')
 const lookupOpInfo = require('./opcodes.js')
 const opFns = require('./opFns.js')
-const constants = require('./constants.js')
+const exceptions = require('./exceptions.js')
 const setImmediate = require('timers').setImmediate
 const BN = utils.BN
 
-const ERROR = constants.ERROR
-const VmError = constants.VmError
+const ERROR = exceptions.ERROR
+const VmError = exceptions.VmError
 
 /**
  * Runs EVM code
